Fix WebGL 1 extension function pointer querying and FULL_ES3
